requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
langchain>=0.2.0
langchain-openai>=0.1.0
gtfs-realtime-bindings==1.0.0
pandas>=1.3.0
numpy>=1.21.0
h5py>=3.0.0
scikit-learn>=0.24.0
pyyaml>=5.4.0
matplotlib>=3.3.0
tensorflow>=2.10.0
torch>=1.12.0
networkx>=2.5.0
